:recycle: Remove duplicate codes and move them into utils
