+++ src/db.rs
@@ -1,4 +1,3 @@
-use std::cmp;
 use std::sync::{mpsc, Arc, Mutex};
 use std::thread;
 
@@ -305,51 +304,6 @@ impl Database {
         Ok(ctx.undefined())
     }
 
-    fn iteration_mode<'a>(
-        options: &options::IterationOption,
-        opt: &'a mut Vec<u8>,
-    ) -> rocksdb::IteratorMode<'a> {
-        let no_range = options.gte.is_none() && options.lte.is_none();
-        if no_range {
-            if options.reverse {
-                rocksdb::IteratorMode::End
-            } else {
-                rocksdb::IteratorMode::Start
-            }
-        } else if options.reverse {
-            *opt = options
-                .lte
-                .clone()
-                .unwrap_or_else(|| vec![255; options.gte.clone().unwrap().len()]);
-            rocksdb::IteratorMode::From(opt, rocksdb::Direction::Reverse)
-        } else {
-            *opt = options
-                .gte
-                .clone()
-                .unwrap_or_else(|| vec![0; options.lte.clone().unwrap().len()]);
-            rocksdb::IteratorMode::From(opt, rocksdb::Direction::Forward)
-        }
-    }
-
-    fn is_breakable(options: &options::IterationOption, key: &[u8], counter: i64) -> bool {
-        if options.limit != -1 && counter >= options.limit {
-            return true;
-        }
-        if options.reverse {
-            if let Some(gte) = &options.gte {
-                if utils::compare(key, gte) == cmp::Ordering::Less {
-                    return true;
-                }
-            }
-        } else if let Some(lte) = &options.lte {
-            if utils::compare(key, lte) == cmp::Ordering::Greater {
-                return true;
-            }
-        }
-
-        false
-    }
-
     pub fn js_iterate(mut ctx: FunctionContext) -> JsResult<JsUndefined> {
         let option_inputs = ctx.argument::<JsObject>(0)?;
         let options = options::IterationOption::new(&mut ctx, option_inputs);
@@ -363,9 +317,9 @@ impl Database {
 
         let a_cb_on_data = Arc::new(Mutex::new(cb_on_data));
         db.send(move |conn, channel| {
-            let iter = conn.iterator(Self::iteration_mode(&options, &mut vec![]));
+            let iter = conn.iterator(utils::iteration_mode(&options, &mut vec![], false));
             for (counter, (key, val)) in iter.enumerate() {
-                if Self::is_breakable(&options, &key, counter as i64) {
+                if utils::is_breakable(&options, &key, counter as i64, false) {
                     break;
                 }
                 let c = a_cb_on_data.clone();
